Rewrite traj to use dist_array and boolean matrices
This replaces the numerous select_atom calls that was slowing down the
rate-limiting step of identifying which waters are in the vicity of
sites.

The new code uses up more memory by calling in_memory versions of only
the atoms that needs their positions accessed frequently, and then
creating a large boolean matrix to store proximity data instead of dicts
of lists of lists.