Ubq branch: added additional analysis metrics
Corrected clustering coefficient to deal with R3 space instead of
arbitrary space.

Added local cohesion as a measure of the rigidity of the local area.

Changed the workflow in 2B to fully load and interpret two comparison
sets. Performance will suffer as a results, dpeending on the total
number of windows.